Move into subdirectory bitblast.
